Move test Relay to its own file
