+++ test/integration/lift_drag_system.cc
@@ -37,7 +37,7 @@
 #include "ignition/gazebo/SystemLoader.hh"
 #include "ignition/gazebo/test_config.hh"
 
-#include "plugins/MockSystem.hh"
+#include "../helpers/Relay.hh"
 
 #define TOL 1e-4
 
@@ -54,45 +54,6 @@ class LiftDragTestFixture : public ::testing::Test
     setenv("IGN_GAZEBO_SYSTEM_PLUGIN_PATH",
            (std::string(PROJECT_BINARY_PATH) + "/lib").c_str(), 1);
   }
-};
-
-class Relay
-{
-  public: Relay()
-  {
-    auto plugin = loader.LoadPlugin("libMockSystem.so",
-                                    "ignition::gazebo::MockSystem", nullptr);
-    EXPECT_TRUE(plugin.has_value());
-
-    this->systemPtr = plugin.value();
-
-    this->mockSystem =
-        dynamic_cast<MockSystem *>(systemPtr->QueryInterface<System>());
-    EXPECT_NE(nullptr, this->mockSystem);
-  }
-
-  public: Relay &OnPreUpdate(MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->preUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay &OnUpdate(MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->updateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay &OnPostUpdate(MockSystem::CallbackTypeConst _cb)
-  {
-    this->mockSystem->postUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: SystemPluginPtr systemPtr;
-
-  private: SystemLoader loader;
-  private: MockSystem *mockSystem;
 };
 
 /////////////////////////////////////////////////
@@ -118,7 +79,7 @@ TEST_F(LiftDragTestFixture, VerifyVerticalForce)
   const std::string jointName = "body_joint";
   const double desiredVel = -0.2;
 
-  Relay testSystem;
+  test::Relay testSystem;
   std::vector<math::Vector3d> linearVelocities;
   std::vector<math::Vector3d> forces;
   testSystem.OnPreUpdate(
@@ -185,7 +146,7 @@ TEST_F(LiftDragTestFixture, VerifyVerticalForce)
   // \todo(addisu) This assumes that the this system will run after the lift
   // drag system. This is needed to capture the wrench set by the lift drag
   // system. This assumption may not hold when systems are run in parallel.
-  Relay wrenchRecorder;
+  test::Relay wrenchRecorder;
   wrenchRecorder.OnPreUpdate([&](const gazebo::UpdateInfo &,
                               const gazebo::EntityComponentManager &_ecm)
       {
