+++ test/integration/log_system.cc
@@ -47,7 +47,7 @@
 #include "ignition/gazebo/SystemLoader.hh"
 #include "ignition/gazebo/test_config.hh"
 
-#include "plugins/MockSystem.hh"
+#include "../helpers/Relay.hh"
 
 using namespace ignition;
 using namespace gazebo;
@@ -149,46 +149,6 @@ void entryDiff(std::vector<std::string> &_paths1,
   _diff.resize(diffIt - _diff.begin());
 }
 #endif
-
-/////////////////////////////////////////////////
-class Relay
-{
-  public: Relay()
-  {
-    auto plugin = loader.LoadPlugin("libMockSystem.so",
-                                    "ignition::gazebo::MockSystem", nullptr);
-    EXPECT_TRUE(plugin.has_value());
-
-    this->systemPtr = plugin.value();
-
-    this->mockSystem =
-        dynamic_cast<MockSystem *>(systemPtr->QueryInterface<System>());
-    EXPECT_NE(nullptr, this->mockSystem);
-  }
-
-  public: Relay &OnPreUpdate(MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->preUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay &OnUpdate(MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->updateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay &OnPostUpdate(MockSystem::CallbackTypeConst _cb)
-  {
-    this->mockSystem->postUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: SystemPluginPtr systemPtr;
-
-  private: SystemLoader loader;
-  private: MockSystem *mockSystem;
-};
 
 //////////////////////////////////////////////////
 class LogSystemTest : public ::testing::Test
@@ -917,7 +877,7 @@ TEST_F(LogSystemTest, LogControl)
 
   Server server(config);
 
-  Relay testSystem;
+  test::Relay testSystem;
   math::Pose3d spherePose;
   bool sphereFound{false};
   testSystem.OnPostUpdate(
@@ -1317,7 +1277,7 @@ TEST_F(LogSystemTest, LogControlLevels)
 
   Server server(config);
 
-  Relay testSystem;
+  test::Relay testSystem;
 
   EntityGraph entityGraph;
