+++ test/integration/physics_system.cc
@@ -55,7 +55,7 @@
 #include "ignition/gazebo/components/Visual.hh"
 #include "ignition/gazebo/components/World.hh"
 
-#include "plugins/MockSystem.hh"
+#include "../helpers/Relay.hh"
 
 using namespace ignition;
 using namespace gazebo;
@@ -71,44 +71,6 @@ class PhysicsSystemFixture : public ::testing::Test
       (std::string(PROJECT_BINARY_PATH) + "/lib").c_str(), 1);
   }
 };
-
-class Relay
-{
-  public: Relay()
-  {
-    auto plugin = sm.LoadPlugin("libMockSystem.so",
-                                "ignition::gazebo::MockSystem",
-                                nullptr);
-    EXPECT_TRUE(plugin.has_value());
-    this->systemPtr = plugin.value();
-    this->mockSystem = static_cast<gazebo::MockSystem *>(
-        systemPtr->QueryInterface<gazebo::System>());
-  }
-
-  public: Relay & OnPreUpdate(gazebo::MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->preUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay & OnUpdate(gazebo::MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->updateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay & OnPostUpdate(gazebo::MockSystem::CallbackTypeConst _cb)
-  {
-    this->mockSystem->postUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: ignition::gazebo::SystemPluginPtr systemPtr;
-
-  private: gazebo::SystemLoader sm;
-  private: gazebo::MockSystem *mockSystem;
-};
-
 
 /////////////////////////////////////////////////
 TEST_F(PhysicsSystemFixture, CreatePhysicsWorld)
@@ -153,7 +115,7 @@ TEST_F(PhysicsSystemFixture, FallingObject)
   std::vector<ignition::math::Pose3d> spherePoses;
 
   // Create a system that records the poses of the sphere
-  Relay testSystem;
+  test::Relay testSystem;
 
   testSystem.OnPostUpdate(
     [modelName, &spherePoses](const gazebo::UpdateInfo &,
@@ -230,7 +192,7 @@ TEST_F(PhysicsSystemFixture, CanonicalLink)
   ASSERT_EQ(3u, expectedLinPoses.size());
 
   // Create a system that records the poses of the links after physics
-  Relay testSystem;
+  test::Relay testSystem;
 
   std::unordered_map<std::string, ignition::math::Pose3d> postUpLinkPoses;
   testSystem.OnPostUpdate(
@@ -293,7 +255,7 @@ TEST_F(PhysicsSystemFixture, RevoluteJoint)
   const std::string modelName{"revolute_demo"};
   const std::string rotatingLinkName{"link2"};
 
-  Relay testSystem;
+  test::Relay testSystem;
 
   std::vector<double> armDistances;
 
@@ -363,7 +325,7 @@ TEST_F(PhysicsSystemFixture, CreateRuntime)
   // create `Relay` systems in the first place. Consider keeping the ECM in a
   // shared pointer owned by the SimulationRunner.
   EntityComponentManager *ecm{nullptr};
-  Relay testSystem;
+  test::Relay testSystem;
   testSystem.OnPreUpdate([&](const gazebo::UpdateInfo &,
                              gazebo::EntityComponentManager &_ecm)
       {
@@ -450,7 +412,7 @@ TEST_F(PhysicsSystemFixture, SetFrictionCoefficient)
   std::map<std::string, std::vector<ignition::math::Pose3d>> poses;
 
   // Create a system that records the poses of the 3 boxes
-  Relay testSystem;
+  test::Relay testSystem;
   double dt = 0.0;
 
   testSystem.OnPostUpdate(
@@ -532,7 +494,7 @@ TEST_F(PhysicsSystemFixture, MultiAxisJointPosition)
 
   server.SetUpdatePeriod(0ns);
 
-  Relay testSystem;
+  test::Relay testSystem;
   // Create JointPosition components if they don't already exist
   testSystem.OnPreUpdate(
       [&](const gazebo::UpdateInfo &, gazebo::EntityComponentManager &_ecm)
@@ -619,7 +581,7 @@ TEST_F(PhysicsSystemFixture, ResetPositionComponent)
 
   const std::string rotatingJointName{"j2"};
 
-  Relay testSystem;
+  test::Relay testSystem;
 
   double pos0 = 0.42;
 
@@ -714,7 +676,7 @@ TEST_F(PhysicsSystemFixture, ResetVelocityComponent)
 
   const std::string rotatingJointName{"j2"};
 
-  Relay testSystem;
+  test::Relay testSystem;
 
   double vel0 = 3.0;
 
@@ -807,7 +769,7 @@ TEST_F(PhysicsSystemFixture, GetBoundingBox)
   std::map<std::string, ignition::math::AxisAlignedBox> bbox;
 
   // Create a system that records the bounding box of a model
-  Relay testSystem;
+  test::Relay testSystem;
 
   testSystem.OnPreUpdate(
     [&](const gazebo::UpdateInfo &,
