+++ test/integration/pose_publisher_system.cc
@@ -31,7 +31,7 @@
 #include "ignition/gazebo/SystemLoader.hh"
 #include "ignition/gazebo/test_config.hh"
 
-#include "plugins/MockSystem.hh"
+#include "../helpers/Relay.hh"
 
 #define tol 10e-4
 
@@ -48,46 +48,6 @@ class PosePublisherTest : public ::testing::TestWithParam<int>
     setenv("IGN_GAZEBO_SYSTEM_PLUGIN_PATH",
            (std::string(PROJECT_BINARY_PATH) + "/lib").c_str(), 1);
   }
-};
-
-class Relay
-{
-  public: Relay()
-  {
-    auto plugin = loader.LoadPlugin("libMockSystem.so",
-                                "ignition::gazebo::MockSystem",
-                                nullptr);
-    EXPECT_TRUE(plugin.has_value());
-
-    this->systemPtr = plugin.value();
-
-    this->mockSystem = static_cast<MockSystem *>(
-        systemPtr->QueryInterface<System>());
-    EXPECT_NE(nullptr, this->mockSystem);
-  }
-
-  public: Relay &OnPreUpdate(MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->preUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay &OnUpdate(MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->updateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay &OnPostUpdate(MockSystem::CallbackTypeConst _cb)
-  {
-    this->mockSystem->postUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: SystemPluginPtr systemPtr;
-
-  private: SystemLoader loader;
-  private: MockSystem *mockSystem;
 };
 
 std::mutex mutex;
@@ -155,7 +115,7 @@ TEST_F(PosePublisherTest, PublishCmd)
   EXPECT_FALSE(*server.Running(0));
 
   // Create a system that records the model's link poses
-  Relay testSystem;
+  test::Relay testSystem;
 
   // The delimiter is hard coded in PosePublisher. If it changes there, it needs
   // to be changed here as well
@@ -366,7 +326,7 @@ TEST_F(PosePublisherTest, UpdateFrequency)
   EXPECT_FALSE(*server.Running(0));
 
   // Create a system that records the model's link poses
-  Relay testSystem;
+  test::Relay testSystem;
 
   {
     std::lock_guard<std::mutex> lock(mutex);
@@ -431,7 +391,7 @@ TEST_F(PosePublisherTest, StaticPosePublisher)
   EXPECT_FALSE(*server.Running(0));
 
   // Create a system that records the model's link poses
-  Relay testSystem;
+  test::Relay testSystem;
 
   {
     std::lock_guard<std::mutex> lock(mutex);
@@ -683,7 +643,7 @@ TEST_F(PosePublisherTest, StaticPoseUpdateFrequency)
   EXPECT_FALSE(*server.Running(0));
 
   // Create a system that records the model's link poses
-  Relay testSystem;
+  test::Relay testSystem;
 
   {
     std::lock_guard<std::mutex> lock(mutex);
