+++ test/integration/user_commands.cc
@@ -32,7 +32,7 @@
 #include "ignition/gazebo/SystemLoader.hh"
 #include "ignition/gazebo/test_config.hh"
 
-#include "plugins/MockSystem.hh"
+#include "../helpers/Relay.hh"
 
 using namespace ignition;
 using namespace gazebo;
@@ -47,47 +47,6 @@ class UserCommandsTest : public ::testing::Test
     setenv("IGN_GAZEBO_SYSTEM_PLUGIN_PATH",
            (std::string(PROJECT_BINARY_PATH) + "/lib").c_str(), 1);
   }
-};
-
-//////////////////////////////////////////////////
-class Relay
-{
-  public: Relay()
-  {
-    auto plugin = loader.LoadPlugin("libMockSystem.so",
-                                    "ignition::gazebo::MockSystem",
-                                    nullptr);
-    EXPECT_TRUE(plugin.has_value());
-
-    this->systemPtr = plugin.value();
-
-    this->mockSystem =
-        dynamic_cast<MockSystem *>(systemPtr->QueryInterface<System>());
-    EXPECT_NE(nullptr, this->mockSystem);
-  }
-
-  public: Relay &OnPreUpdate(MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->preUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay &OnUpdate(MockSystem::CallbackType _cb)
-  {
-    this->mockSystem->updateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: Relay &OnPostUpdate(MockSystem::CallbackTypeConst _cb)
-  {
-    this->mockSystem->postUpdateCallback = std::move(_cb);
-    return *this;
-  }
-
-  public: SystemPluginPtr systemPtr;
-
-  private: SystemLoader loader;
-  private: MockSystem *mockSystem;
 };
 
 /////////////////////////////////////////////////
@@ -109,7 +68,7 @@ TEST_F(UserCommandsTest, Create)
   // create `Relay` systems in the first place. Consider keeping the ECM in a
   // shared pointer owned by the SimulationRunner.
   EntityComponentManager *ecm{nullptr};
-  Relay testSystem;
+  test::Relay testSystem;
   testSystem.OnPreUpdate([&](const gazebo::UpdateInfo &,
                              gazebo::EntityComponentManager &_ecm)
       {
@@ -371,7 +330,7 @@ TEST_F(UserCommandsTest, Remove)
   // create `Relay` systems in the first place. Consider keeping the ECM in a
   // shared pointer owned by the SimulationRunner.
   EntityComponentManager *ecm{nullptr};
-  Relay testSystem;
+  test::Relay testSystem;
   testSystem.OnPreUpdate([&](const gazebo::UpdateInfo &,
                              gazebo::EntityComponentManager &_ecm)
       {
@@ -555,7 +514,7 @@ TEST_F(UserCommandsTest, Pose)
 
   // Create a system just to get the ECM
   EntityComponentManager *ecm{nullptr};
-  Relay testSystem;
+  test::Relay testSystem;
   testSystem.OnPreUpdate([&](const gazebo::UpdateInfo &,
                              gazebo::EntityComponentManager &_ecm)
       {
